fix: gitignore .env 적용 및 editorpage 적용
