+++ src/pages/CodeEditorPage.tsx
@@ -2,7 +2,6 @@ import { useEffect, useRef, useState } from 'react';
 import Sidebar from '../components/codeeditorpage/Sidebar';
 import Header from '../components/codeeditorpage/Header';
 import SubHeader from '../components/codeeditorpage/SubHeader';
-import MonacoEditor from '../components/codeeditorpage/MonacoEditor';
 import type { FileNode } from '../types/CodeEditor.types';
 import ConsoleBox from '../components/codeeditorpage/ConsoleBox';
 import ProblemPreview from '../components/codeeditorpage/ProblemPreview';
@@ -30,19 +29,68 @@ import { getRandomColor, getRandomName } from '../utils/cursor';
 
 // 인터페이스 설정
 interface Message {
+  nickname: string;
   time: string;
-  sender: 'me' | 'other';
-  text: string;
-  name: string;
+  content: string;
 }
 
-export default function CodeEditorPage() {
+// 소켓 관련 핸들러 저장용 Map
+const socketHandlers = new Map<
+  string,
+  {
+    handleSync: (u: Uint8Array) => void;
+    handleUpdate: (u: Uint8Array) => void;
+    handleAwarenessUpdateToSocket: ({
+      update,
+      fileName,
+    }: {
+      update: Uint8Array;
+      fileName: string;
+    }) => void;
+
+    handleChatUpdate: (msg: Message) => void;
+
+    handleChatSync: (msg: Message[]) => void;
+  }
+>();
+
+// Y.Doc 업데이트 핸들러 저장 Map
+const ydocHandlers = new Map<
+  string,
+  {
+    handleDocUpdate: (u: Uint8Array) => void;
+  }
+>();
+
+// Awareness 이벤트 핸들러 저장 Map
+const awarenessHandlers = new Map<
+  string,
+  {
+    handleAwarenessUpdate: (e: any) => void;
+    handleAwarenessChange: () => void;
+  }
+>();
+
+const ydocs: Record<string, Y.Doc> = {}; // 파일별 Y.Doc 인스턴스를 저장
+const awarenessMap: Record<string, awarenessProtocol.Awareness> = {}; // 파일별 Awareness 인스턴스를 저장
+const bindingMap: Record<string, MonacoBinding> = {}; // 파일별 MonacoBinding 인스턴스를 저장
+const modelMap: Record<string, monaco.editor.ITextModel> = {}; // 파일별 Monaco 모델 저장
+
+const username = getRandomName(); // 사용자 이름 생성 (나중에 변경해야 함)
+const usercolor = getRandomColor(); // 사용자 커서 색상 지정
+
+export default function CodeEditorPage3() {
   const [isPreviewVisible, setIsPreviewVisible] = useState<boolean>(false);
   const [isDarkMode, setIsDarkMode] = useState<boolean>(false);
   const [activePanel, setActivePanel] = useState<'chat' | 'ai' | null>(null);
   const [inviteModalOpen, setInviteModalOpen] = useState(false);
   const [settingModalOpen, setSettingModalOpen] = useState(false);
-  const [currentFile, setCurrentFile] = useState<string>('');
+  const cursorListenerRef = useRef<monaco.IDisposable | null>(null);
+
+  let lastCursorState: { line: number; column: number } = {
+    line: 0,
+    column: 0,
+  };
 
   const dummyFileTree: FileNode[] = [
     {
@@ -52,14 +100,17 @@ export default function CodeEditorPage() {
       childrenNode: [
         {
           id: '2',
-          name: 'App.tsx',
+          name: 'index.html',
           type: 'file',
         },
         {
           id: '3',
           name: 'components',
           type: 'folder',
-          childrenNode: [{ id: '4', name: 'Sidebar.tsx', type: 'file' }],
+          childrenNode: [
+            { id: '4', name: 'style.css', type: 'file' },
+            { id: '5', name: 'app.js', type: 'file' },
+          ],
         },
       ],
     },
@@ -69,18 +120,9 @@ export default function CodeEditorPage() {
     { id: 2, nickname: '수연', line: 13 },
     { id: 3, nickname: '예람', line: 87 },
   ];
-  const filename = 'Main.tsx';
+
   const isOwner = true;
   const mode = 'problem';
-  const chatMessages = [
-    {
-      nickname: '기영',
-      time: '10:15',
-      content: '여기 문제 조건 다시 확인해주세요.',
-    },
-    { nickname: '민수', time: '10:17', content: '네 알겠습니다!' },
-    { nickname: '하린', time: '10:18', content: '힌트는 어디에 있나요?' },
-  ];
 
   const aiMessages = [
     {
@@ -101,66 +143,59 @@ export default function CodeEditorPage() {
   const { roomId, type } = location.state || {};
 
   // 모나코, yjs, awareness, socket io
-  const containerRef = useRef<HTMLDivElement>(null); // 에디터가 들어갈 div
   const monacoRef = useRef<monaco.editor.IStandaloneCodeEditor | null>(null); // Monaco 인스턴스
   const socketRef = useRef<Socket | null>(null);
-  const awarenessRef = useRef<awarenessProtocol.Awareness | null>(null);
-
-  const ydocRef = useRef<Y.Doc | null>(null);
   const oldDecorationsRef = useRef<string[]>([]);
 
-  const [messages, setMessages] = useState<Message[]>([]);
+  const containerRef = useRef<HTMLDivElement>(null);
+  const currentFileRef = useRef<string>('index.html');
+  const [oldFile, setOldFile] = useState<string>('');
+
+  const [messages, setMessages] = useState<Message[]>([
+    {
+      nickname: '기영',
+      time: '10:15',
+      content: '여기 문제 조건 다시 확인해주세요.',
+    },
+    {
+      nickname: '민수',
+      time: '10:17',
+      content: '네 알겠습니다!',
+    },
+    {
+      nickname: '하린',
+      time: '10:18',
+      content: '힌트는 어디에 있나요?',
+    },
+  ]);
+
   const [input, setInput] = useState('');
 
   // 추가 부분
   useEffect(() => {
-    // ✅ 소켓이 생성되지 않았다면 최초 1회만 생성
     if (!socketRef.current) {
-      socketRef.current = initSocket();
+      socketRef.current = initSocket(); // 소켓 초기화
     }
     const socket = socketRef.current;
 
-    // ✅ Y.Doc 인스턴스도 최초 1회만 생성
-    if (!ydocRef.current) {
-      ydocRef.current = new Y.Doc();
-    }
+    socket.on('connect', () => {
+      console.log('🔌 Connected to server', socket.id);
+      socket.emit('join', { roomId }); // 방 참가 요청
+    });
 
-    const ydoc = ydocRef.current;
-    const ytext = ydoc.getText('shared');
+    // 채팅 동기화 요청
+    socket.emit('chat-sync', { roomId });
 
-    // ✅ Awareness 객체도 1회만 생성하고 사용자 상태 등록
-    if (!awarenessRef.current) {
-      awarenessRef.current = initAwareness(ydoc);
-    }
-    const awareness = awarenessRef.current;
-
-    // 🔁 실시간 문서 동기화 관련 소켓 이벤트 설정
-    socketEventsEdit(socket, ydoc, awareness);
-
-    // 💬 채팅 관련 소켓 이벤트 설정
-    socketEventChat(awareness);
-
-    // 🖱️ 다른 사용자 커서 위치 추적 및 표시
-    awarenessEvents(awareness, socket);
-
-    // ⚠️ Monaco 에디터를 렌더링할 DOM이 없다면 중단
-    if (!containerRef.current) return;
-
-    // 📝 Monaco 에디터 생성 및 Awareness에 커서 위치 반영 설정
-    const editor = initEditor(containerRef.current, awareness);
-    monacoRef.current = editor;
-
-    // 🔗 Monaco 에디터와 Yjs 문서 및 awareness 연결
-    const binding = new MonacoBinding(
-      ytext,
-      editor.getModel()!,
-      new Set([editor]),
-      awareness,
-    );
+    // 초기 파일에 대한 Y.Doc 초기화
+    initYDocIfNeeded(currentFileRef.current, socket);
 
     // ✅ 브라우저 종료 시 사용자 상태 정리
     const cleanUp = () => {
-      awareness.setLocalState(null);
+      // 페이지의 awareness값 삭제
+      const oldAwareness = awarenessMap[currentFileRef.current];
+      oldAwareness.setLocalState(null); // 내 커서 상태 제거
+      const socket = socketRef.current; // 해당 클라이언트의 awareness값을 삭제
+      socket?.emit('awareness-remove', { roomId }); // 서버에 상태 제거 요청
     };
 
     window.addEventListener('beforeunload', cleanUp);
@@ -169,158 +204,423 @@ export default function CodeEditorPage() {
     return () => {
       cleanUp();
       window.removeEventListener('beforeunload', cleanUp);
-      awareness.destroy();
-      binding.destroy();
-      editor.dispose();
-      awarenessRef.current?.destroy?.();
-      monacoRef.current?.dispose?.();
-      socketRef.current?.disconnect?.();
-      awarenessRef.current = null;
-      monacoRef.current = null;
-      socketRef.current = null;
       setTimeout(() => socket.disconnect(), 50); // 💡 100~200ms 정도 기다려주는 게 일반적
     };
-  }, [roomId]);
+  }, []);
 
   // ✅ 소켓 초기화 함수 (Fast Refresh 대응 없이 단순 생성)
   const initSocket = () => {
-    return io('http://localhost:3002/collab-algorism', {
+    //https://codetogether.store/collab-webpublish
+    //http://localhost:3002/collab-webpublish
+    return io('https://codetogether.store/collab-webpublish', {
       transports: ['websocket'],
       withCredentials: true,
     });
   };
 
-  // ✅ Awareness 초기화 및 사용자 정보 등록
-  const initAwareness = (ydoc: Y.Doc) => {
+  // 👉 유틸: Y.Doc + Awareness 생성
+  const createYDocAndAwareness = (fileName: string) => {
+    const ydoc = new Y.Doc();
+    ydocs[fileName] = ydoc;
     const awareness = new awarenessProtocol.Awareness(ydoc);
-
-    awareness.setLocalStateField('user', {
-      name: getRandomName(),
-      color: getRandomColor(),
-    });
-    return awareness;
+    awareness.setLocalStateField('user', { name: username, color: usercolor });
+    awarenessMap[fileName] = awareness;
+    return { ydoc, awareness };
+  };
+  // 👉 유틸: Monaco 모델 생성
+  const createMonacoModel = (fileName: string, ytext: Y.Text) => {
+    const uri = monaco.Uri.parse(`file:///${fileName}`);
+    const existingModel = monaco.editor.getModel(uri);
+    if (existingModel) existingModel.dispose();
+    const model = monaco.editor.createModel(
+      ytext.toString(),
+      getLanguageFromFile(fileName),
+      uri,
+    );
+    modelMap[fileName] = model;
+    return model;
   };
 
-  // 📡 문서 편집과 관련된 소켓 이벤트 설정
-  //socket.on(...)에 등록하는 콜백은 어디서 선언되었는지는 중요하지 않음
-  // 해당 시점에 유효한 함수라면 어떤 함수든 가능
-  const socketEventsEdit = (
-    socket: Socket,
-    ydoc: Y.Doc,
-    awareness: awarenessProtocol.Awareness,
-  ) => {
-    socket.on('connect', () => {
-      console.log('🔌 Connected to server', socket.id);
-      console.log('roomId:', roomId);
+  // ✅ Yjs 초기화 필요 시만 생성
+  const initYDocIfNeeded = (fileName: string, socket: Socket) => {
+    const { ydoc, awareness } = createYDocAndAwareness(fileName);
 
-      socket.emit('join', { roomId });
+    const ytext = ydoc.getText(fileName);
+    const model = createMonacoModel(fileName, ytext);
+
+    // 에디터 레퍼런스 생성
+    monacoRef.current = monaco.editor.create(containerRef.current!, {
+      model,
+      theme: 'vs-dark',
+      automaticLayout: true,
     });
 
-    // 서버로부터 문서 초기 상태 동기화
-    socket.on('sync', (update: Uint8Array) => {
-      const fixedUpdate =
-        update instanceof Uint8Array ? update : new Uint8Array(update);
-      if (fixedUpdate instanceof Uint8Array && fixedUpdate.length > 0) {
-        Y.applyUpdate(ydoc, fixedUpdate);
+    // 👆 커서 리스너 중복 제거 후 등록
+    if (cursorListenerRef.current) cursorListenerRef.current.dispose();
+
+    // 커서 변경 시 이벤트 발동
+    cursorListenerRef.current = monacoRef.current.onDidChangeCursorPosition(
+      (e) => {
+        const current = awareness.getLocalState()?.cursor;
+        const next = { line: e.position.lineNumber, column: e.position.column };
+        if (
+          !current ||
+          current.line !== next.line ||
+          current.column !== next.column
+        ) {
+          awareness.setLocalStateField('cursor', next);
+        }
+      },
+    );
+
+    // 🔗 바인딩 연결
+    const binding = new MonacoBinding(
+      ytext,
+      model,
+      new Set([monacoRef.current]),
+      awareness,
+    );
+
+    // 바인딩 저장
+    bindingMap[fileName] = binding;
+
+    registerSocketEvents(fileName, socket);
+  };
+
+  const registerSocketEvents = (fileName: string, socket: Socket) => {
+    // 기존 이벤트 삭제
+    const ydoc = ydocs[oldFile]!;
+    const awareness = awarenessMap[oldFile]!;
+
+    console.log('삭제할 이벤트 파일: ', oldFile);
+
+    // 🔁 기존 핸들러 제거
+    const prevSocketHandler = socketHandlers.get(oldFile);
+
+    if (prevSocketHandler) {
+      socket.off('sync', prevSocketHandler.handleSync);
+      socket.off('update', prevSocketHandler.handleUpdate);
+      socket.off(
+        'awareness-update',
+        prevSocketHandler.handleAwarenessUpdateToSocket,
+      );
+      socket.off('chat', prevSocketHandler.handleChatUpdate);
+      socket.off('chat', prevSocketHandler.handleChatSync);
+      console.log('prevSocketHandler 이벤트 제거');
+    }
+
+    const prevYdocHandler = ydocHandlers.get(oldFile);
+    if (prevYdocHandler) {
+      ydoc.off('update', prevYdocHandler.handleDocUpdate);
+      console.log('prevYdocHandler 이벤트 제거');
+    }
+
+    const prevAwarenessHandler = awarenessHandlers.get(oldFile);
+    if (prevAwarenessHandler) {
+      awareness.off('update', prevAwarenessHandler.handleAwarenessUpdate);
+      awareness.off('change', prevAwarenessHandler.handleAwarenessChange);
+
+      console.log('prevAwarenessHandler 이벤트 제거');
+    }
+
+    // 새 Ydoc 와 Awareness값 가져오기
+    const newYdoc = ydocs[fileName]!;
+    const newAwareness = awarenessMap[fileName]!;
+
+    // 🔧 새로운 핸들러 정의 및 등록
+
+    // Ydoc 문서 동기화 핸들러
+    const handleSync = (update: Uint8Array) => {
+      if (!newYdoc) {
+        console.log('설마 없어?');
       }
-    });
+      console.log('Ydoc 동기화:', update);
+      Y.applyUpdate(newYdoc, new Uint8Array(update));
+    };
 
-    // 다른 클라이언트의 업데이트 반영
-    socket.on('update', (update: Uint8Array | ArrayBuffer) => {
-      console.log('update:', update);
+    // Ydoc 문서 Update
+    const handleUpdate = (update: Uint8Array) => {
+      Y.applyUpdate(newYdoc, new Uint8Array(update));
+    };
 
-      try {
-        const fixedUpdate =
-          update instanceof Uint8Array ? update : new Uint8Array(update);
-        Y.applyUpdate(ydoc, new Uint8Array(fixedUpdate));
-      } catch (err) {
-        console.error('update apply 실패:', err);
+    // Awareness 상태 Update
+    const handleAwarenessUpdateToSocket = ({
+      update,
+      fileName,
+    }: {
+      update: Uint8Array;
+      fileName: string;
+    }) => {
+      console.log(
+        `awareness update - fileName:${fileName} - currentFile:${currentFileRef.current}`,
+      );
+
+      // 내 페이지 있는 커서만 보여주기
+      if (fileName !== currentFileRef.current) {
+        console.log(
+          '다른 파일 awareness 업데이터:',
+          fileName,
+          currentFileRef.current,
+        );
+        return;
       }
-    });
 
-    // 현재 클라이언트의 업데이트를 서버로 전송
-    ydoc.on('update', (update: Uint8Array) => {
-      socket.emit('update', { roomId, update });
-    });
-
-    // Awareness 상태 동기화
-    socket.on('awareness-update', (update: Uint8Array) => {
       awarenessProtocol.applyAwarenessUpdate(
-        awareness,
+        newAwareness,
         new Uint8Array(update),
         null,
       );
+    };
+
+    // 채팅 업데이트
+    const handleChatUpdate = (newMessage: Message) => {
+      console.log('receivedMessage:', newMessage);
+      setMessages((prev) => [...prev, newMessage]);
+    };
+
+    // 채팅 동기화
+    const handleChatSync = (messages: Message[]) => {
+      console.log('동기화 테스트', messages);
+      if (messages.length == 0) {
+        console.log('동기화 메시지 빈값');
+      }
+      messages.map((message) => {
+        console.log(message.content);
+        setMessages((prev) => [...prev, message]);
+      });
+    };
+
+    // 이벤트 등록
+    socket.on('sync', handleSync);
+    socket.on('update', handleUpdate);
+    socket.on('awareness-update', handleAwarenessUpdateToSocket);
+    socket.on('chat', handleChatUpdate);
+    socket.on('chat-sync', handleChatSync);
+
+    // 파일 싱크 요청
+    socket.emit('sync', { roomId, fileName });
+
+    // 핸들러 레퍼런스 저장
+    socketHandlers.set(fileName, {
+      handleSync,
+      handleUpdate,
+      handleAwarenessUpdateToSocket,
+      handleChatUpdate,
+      handleChatSync,
     });
 
-    // 💬 채팅 메시지 수신 처리
-    socket.on('chat', (receivedMessage: Message) => {
-      console.log('receivedMessage:', receivedMessage);
-      setMessages((prev) => [...prev, receivedMessage]);
-    });
-  };
+    // 기존 이벤트 핸들러 정의 (항상 같은 참조 사용)
+    const handleDocUpdate = (update: Uint8Array) => {
+      console.log('📄 Y.Doc 업데이트 발생:', fileName);
+      socket.emit('update', { roomId, fileName, update });
+    };
 
-  // 💬 사용자 Awareness에서 유저 이름 가져오기
-  const socketEventChat = (awareness: awarenessProtocol.Awareness) => {
-    const state = awareness.getLocalState();
-    if (!state) return;
+    // 👉  Ydoc 해들러 등록
+    newYdoc.on('update', handleDocUpdate);
 
-    const userName = state.user.name;
-    console.log(userName);
+    // 핸들러 저장
+    ydocHandlers.set(fileName, { handleDocUpdate });
 
-    // socket.emit("update", { roomId, update });
-  };
+    // Awareness 업데이트 핸들러
+    const handleAwarenessUpdate = ({ added, updated, removed }: any) => {
+      console.log('awareness update');
 
-  // 🎯 다른 사용자 Awareness 정보 기반 커서 표시 이벤트
-  const awarenessEvents = (
-    awareness: awarenessProtocol.Awareness,
-    socket: Socket,
-  ) => {
-    // Awareness 업데이트가 발생하면 서버에 전송
-    awareness.on('update', ({ added, updated, removed }) => {
-      const update = awarenessProtocol.encodeAwarenessUpdate(
-        awareness,
-        added.concat(updated, removed),
-      );
-      socket.emit('awareness-update', { roomId, update });
-    });
-
-    // Awareness 상태가 변경되면 커서 표시 업데이트
-    awareness.on('change', () => {
+      const awareness = awarenessMap[fileName];
       const states = awareness.getStates();
-      const newDecorations: monaco.editor.IModelDeltaDecoration[] = [];
+      const myState = states.get(awareness.clientID);
+      const currentCursor = myState?.cursor;
 
-      for (const [clientId, state] of states) {
-        if (clientId === awareness.clientID) continue; // 본인 제외
-        const user = state.user;
-        const cursor = state.cursor;
-        if (!user || !cursor) continue; // 정보 없으면 생략
+      console.log('👤 awareness 업데이트 발생');
 
-        newDecorations.push({
-          range: new monaco.Range(
-            cursor.line,
-            cursor.column,
-            cursor.line,
-            cursor.column,
-          ),
-          options: {
-            className: 'remote-cursor',
-            afterContentClassName: `remote-cursor-label-${clientId}`,
-            stickiness:
-              monaco.editor.TrackedRangeStickiness.NeverGrowsWhenTypingAtEdges,
-          },
-        });
+      console.log('현재 파일:', currentFileRef.current);
+      if (
+        currentCursor &&
+        currentCursor.line === lastCursorState.line &&
+        currentCursor.column === lastCursorState.column
+      ) {
+        console.log(
+          'currentCursor.line:',
+          currentCursor.line,
+          'currentCursor.column:',
+          currentCursor.column,
+        );
+        console.log(
+          'lastCursorState.line:',
+          lastCursorState.line,
+          'lastCursorState.column:',
+          lastCursorState.column,
+        );
 
-        // 🏷️ 유저 커서 이름 스타일 DOM에 추가
-        addCursorLabelStyle(clientId, user.name, user.color);
+        console.log('변동 없음');
+        return;
       }
 
-      const editor = monacoRef.current!;
-      const decorationIds = editor.deltaDecorations(
-        oldDecorationsRef.current,
-        newDecorations,
+      const update = awarenessProtocol.encodeAwarenessUpdate(
+        newAwareness,
+        added.concat(updated, removed),
       );
-      oldDecorationsRef.current = decorationIds;
+      socket.emit('awareness-update', { roomId, fileName, update });
+      console.log('awareness broadcast');
+      // 상태 업데이트 진행
+      lastCursorState = currentCursor ?? null;
+    };
+
+    // 상태 변화 업데이트 감지
+    const handleAwarenessChange = () => {
+      console.log('change 이벤트 발생');
+      updateRemoteCursors(newAwareness);
+    };
+
+    newAwareness.on('update', handleAwarenessUpdate);
+    newAwareness.on('change', handleAwarenessChange);
+
+    awarenessHandlers.set(fileName, {
+      handleAwarenessUpdate,
+      handleAwarenessChange,
     });
+
+    //oldfile 전환
+    setOldFile(fileName);
+  };
+
+  // ✅ 파일 전환 처리
+  const switchToFile = (fileName: string) => {
+    // 기존 파일에 있던 awareness 값 비활성화
+    const oldAwareness = awarenessMap[currentFileRef.current];
+    oldAwareness.setLocalState(null);
+
+    const socket = socketRef.current!;
+
+    // 🔧 이전 바인딩
+    if (bindingMap[oldFile]) {
+      console.log('binding:', oldFile);
+      bindingMap[oldFile].destroy();
+      delete bindingMap[oldFile];
+    }
+
+    // 이전  모델 제거
+    if (modelMap[oldFile]) {
+      console.log('model:', oldFile);
+      const model = modelMap[oldFile];
+      if (!model.isDisposed()) model.dispose();
+      delete modelMap[oldFile];
+    }
+
+    // ✅ 기존 ydoc, awareness는 무조건 파괴하고 새로 생성 (기존 상태 꼬임 방지)
+    if (ydocs[fileName]) {
+      ydocs[fileName].destroy();
+      delete ydocs[fileName];
+    }
+    if (awarenessMap[fileName]) {
+      awarenessMap[fileName].destroy();
+      delete awarenessMap[fileName];
+    }
+
+    // ✅ 새 ydoc / awareness 생성 및 저장
+    const { ydoc, awareness } = createYDocAndAwareness(fileName);
+    ydocs[fileName] = ydoc;
+    awarenessMap[fileName] = awareness;
+
+    // 🔧 모델이 없으면 새로 생성
+    const ytext = ydoc.getText(fileName);
+    let model = modelMap[fileName];
+    if (!model) model = createMonacoModel(fileName, ytext);
+
+    // 🖋️ Monaco 에디터 모델 적용
+    if (!monacoRef.current || monacoRef.current.getModel() === null) {
+      // 기존 인스턴스가 있다면 먼저 dispose
+      if (monacoRef.current) {
+        console.log('기존 코드 모델 삭제');
+        monacoRef.current.dispose();
+        monacoRef.current = null;
+      }
+
+      monacoRef.current = monaco.editor.create(containerRef.current!, {
+        model,
+        theme: 'vs-dark',
+        automaticLayout: true,
+      });
+    } else {
+      monacoRef.current.setModel(model);
+    }
+
+    // 👆 커서 리스너 중복 제거 후 등록
+    if (cursorListenerRef.current) cursorListenerRef.current.dispose();
+    cursorListenerRef.current = monacoRef.current.onDidChangeCursorPosition(
+      (e) => {
+        const current = awareness.getLocalState()?.cursor;
+        const next = { line: e.position.lineNumber, column: e.position.column };
+        if (
+          !current ||
+          current.line !== next.line ||
+          current.column !== next.column
+        ) {
+          awareness.setLocalStateField('cursor', next);
+        }
+      },
+    );
+
+    // 🔗 바인딩 연결
+    const binding = new MonacoBinding(
+      ytext,
+      model,
+      new Set([monacoRef.current]),
+      awareness,
+    );
+    bindingMap[fileName] = binding;
+
+    // 📌 이벤트 재등록
+    registerSocketEvents(fileName, socket);
+  };
+
+  /** ✅ Awareness 기반 사용자 커서 표시 */
+  const updateRemoteCursors = (awareness: awarenessProtocol.Awareness) => {
+    const editor = monacoRef.current!;
+    if (!editor) {
+      console.warn('[awareness] editor가 아직 생성되지 않음');
+      return;
+    }
+
+    const states = awareness.getStates();
+    const newDecorations: monaco.editor.IModelDeltaDecoration[] = [];
+
+    for (const [clientId, state] of states) {
+      if (clientId === awareness.clientID) continue;
+      const user = state.user;
+      const cursor = state.cursor;
+      if (!user || !cursor) continue;
+
+      newDecorations.push({
+        range: new monaco.Range(
+          cursor.line,
+          cursor.column,
+          cursor.line,
+          cursor.column,
+        ),
+        options: {
+          className: 'remote-cursor',
+          afterContentClassName: `remote-cursor-label-${clientId}`,
+          stickiness:
+            monaco.editor.TrackedRangeStickiness.NeverGrowsWhenTypingAtEdges,
+        },
+      });
+
+      addCursorLabelStyle(clientId, user.name, user.color);
+    }
+
+    const decorationIds = editor.deltaDecorations(
+      oldDecorationsRef.current,
+      newDecorations,
+    );
+    oldDecorationsRef.current = decorationIds;
+  };
+
+  /** ✅ 파일 이름에 따라 언어 결정 */
+  const getLanguageFromFile = (filename: string): string => {
+    if (filename.endsWith('.html')) return 'html';
+    if (filename.endsWith('.css')) return 'css';
+    if (filename.endsWith('.js')) return 'javascript';
+    return 'plaintext';
   };
 
   const addCursorLabelStyle = (
@@ -345,57 +645,6 @@ export default function CodeEditorPage() {
     document.head.appendChild(style);
   };
 
-  // 📝 Monaco Editor 초기화 및 Awareness 커서 반영
-  const initEditor = (
-    container: HTMLElement,
-    awareness: awarenessProtocol.Awareness,
-  ): monaco.editor.IStandaloneCodeEditor => {
-    const editor = monaco.editor.create(container, {
-      value: '',
-      language: 'javascript',
-      theme: 'vs-white',
-      automaticLayout: true,
-    });
-
-    editor.onDidChangeCursorPosition((e) => {
-      awareness.setLocalStateField('cursor', {
-        line: e.position.lineNumber,
-        column: e.position.column,
-      });
-    });
-
-    return editor;
-  };
-
-  // ✉️ 채팅 메시지 전송 함수
-  const sendMessage = () => {
-    if (input.trim() === '') return;
-    if (!awarenessRef.current) return;
-
-    const state = awarenessRef.current.getLocalState();
-
-    if (!state) return;
-
-    const userName = state.user.name;
-
-    const localtime = new Date().toLocaleTimeString();
-
-    const message: Message = {
-      time: localtime,
-      sender: 'me',
-      text: input,
-      name: '',
-    };
-
-    setMessages((prev) => [...prev, message]);
-    setInput('');
-
-    const socket = socketRef.current;
-    if (!socket) return;
-    const newMessage = { ...message, name: userName };
-    socket.emit('chat', { roomId, newMessage });
-  };
-
   const problemRef = useRef<HTMLTextAreaElement>(null);
   const chatInputRef = useRef<HTMLInputElement>(null);
 
@@ -414,24 +663,46 @@ export default function CodeEditorPage() {
     console.log('실행');
   };
 
+  // ✉️ 채팅 메시지 전송 함수
   const handleSendChat = () => {
     console.log('채팅:', chatInputRef.current?.value);
+    const content = chatInputRef.current?.value;
+
+    if (!content) return;
+
+    const userName = username;
+    const localtime = new Date().toLocaleTimeString();
+    const newMessage: Message = {
+      time: localtime,
+      content,
+      nickname: userName,
+    };
+    setMessages((prev) => [...prev, newMessage]);
+    setInput('');
+    const socket = socketRef.current;
+    if (!socket) return;
+    socket.emit('chat', { roomId, newMessage });
   };
 
   const handleSelectFile = (file: { fileId: string; filename: string }) => {
     const { fileId, filename } = file;
     console.log(fileId, filename);
 
-    if (filename !== currentFile) {
-      setCurrentFile(filename);
+    if (filename !== currentFileRef.current || !ydocs[filename]) {
       console.log('파일 변경:', filename);
+
+      if (socketRef.current) {
+        console.log('파일변경');
+        switchToFile(filename);
+        currentFileRef.current = filename;
+      }
     }
   };
 
   return (
     <div className="flex flex-col items-center justify-start w-screen h-screen bg-gray-100 dark:bg-black text-black dark:text-white transition-colors min-w-0">
       <Header
-        filename={filename}
+        filename={currentFileRef.current}
         isOwner={isOwner}
         onToggleDarkMode={handleThemeToggle}
         onSettingClick={() => setSettingModalOpen(true)}
@@ -442,7 +713,7 @@ export default function CodeEditorPage() {
           users={users}
           className="border border-gray-200 dark:border-gray-700"
           onInviteClick={() => setInviteModalOpen(true)}
-          currentFile={currentFile}
+          currentFile={currentFileRef.current}
           onSelectFile={handleSelectFile}
         />
         <div className="flex flex-col flex-grow shrink min-w-0">
@@ -454,8 +725,6 @@ export default function CodeEditorPage() {
           />
           <div className="flex w-full h-full">
             <div className={isPreviewVisible ? 'w-1/2' : 'w-full'}>
-              {/* <MonacoEditor theme={isDarkMode ? 'dark' : 'light'} /> */}
-
               {/* 모나코 에디터 */}
               <div
                 ref={containerRef}
@@ -511,10 +780,10 @@ export default function CodeEditorPage() {
             />
             <ChatPanel
               activePanel={activePanel}
-              chatMessages={chatMessages}
+              chatMessages={messages}
               aiMessages={aiMessages}
               inputRef={chatInputRef}
-              onSendChat={sendMessage}
+              onSendChat={handleSendChat}
             />
           </div>
         )}
